Update lisk-validator to contain lisk meta schema
