+++ elements/lisk-validator/package.json
@@ -41,7 +41,8 @@
 		"@types/validator": "12.0.1",
 		"ajv": "6.12.0",
 		"semver": "7.1.3",
-		"validator": "12.2.0"
+		"validator": "12.2.0",
+		"debug": "4.1.1"
 	},
 	"devDependencies": {
 		"@types/jest": "25.1.3",
@@ -56,6 +57,7 @@
 		"jest": "25.1.0",
 		"jest-extended": "0.11.5",
 		"jest-when": "2.7.0",
+		"lodash.clonedeep": "4.5.0",
 		"prettier": "1.19.1",
 		"source-map-support": "0.5.16",
 		"ts-jest": "25.2.1",
