+++ elements/lisk-validator/src/index.ts
@@ -12,7 +12,8 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-import { validator, ErrorObject } from './lisk_validator';
+import { validator, liskSchemaIdentifier } from './lisk_validator';
 
 export * from './validation';
-export { validator, ErrorObject };
+export * from './errors';
+export { validator, liskSchemaIdentifier };
